Final version of the Fire & Smoke Detection Model
This script contains the implementation of a Fire & Smoke detection system trained on over 1000 images divided into 3 categories: "Fire", "Smoke" and "Neutral" meaning it does not showcase any fire, nor smoke. The architectures used were the Resnet50 and Xception; both of them were used as feature extractors, putting on top of the 2 extra trainable layers to improve accuracy.  After only 11 minutes of training the reached accuracy was of 0.3085 and 0.3264 for the training and validation sets, respectively. Further training and custom transformations are yet to be performed in order to try to improve accuracy. More performance metrics should also be computed for increased reliability.